# MINOR: optimize EvictableKey/LastUsedKey compareTo function