The previous implementation create two object wraps. 

1. tuple
2. primitive type -> object type

As ```compareTo``` is hot method in fetch path, it should be worth rewriting them by java code to avoid useless wrap.

**BEFORE**
```
FetchKeyBenchmark.compareEvictableKey                               avgt   15    12.156 ±   0.010   ns/op
FetchKeyBenchmark.compareEvictableKey:·gc.alloc.rate.norm           avgt   15   112.000 ±   0.001    B/op
FetchKeyBenchmark.compareLastUsedKey                                avgt   15    13.929 ±   0.088   ns/op
FetchKeyBenchmark.compareLastUsedKey:·gc.alloc.rate.norm            avgt   15   136.000 ±   0.001    B/op
```

**AFTER**
```
FetchKeyBenchmark.compareEvictableKey                      avgt   15   2.351 ±  0.001   ns/op
FetchKeyBenchmark.compareEvictableKey:·gc.alloc.rate.norm  avgt   15  ≈ 10⁻⁷             B/op
FetchKeyBenchmark.compareLastUsedKey                       avgt   15   2.154 ±  0.001   ns/op
FetchKeyBenchmark.compareLastUsedKey:·gc.alloc.rate.norm   avgt   15  ≈ 10⁻⁷             B/op
```

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)